avoid circular import for decorator
