+++ fossunited/utils/decorators.py
@@ -12,12 +12,13 @@ from fossunited.api.chapter import (
     check_if_chapter_or_event_core_member,
     check_if_event_member,
 )
-from fossunited.api.hackathon import is_localhost_organizer
 from fossunited.doctype_ids import (
     CAMPAIGN,
     HACKATHON_LOCALHOST,
     HACKATHON_PARTICIPANT,
     HACKATHON_TEAM,
+    LOCALHOST_ORGANIZER,
+    USER_PROFILE,
 )
 
 
@@ -195,6 +196,38 @@ def require_chapter_or_event_member(event_id="event"):
         return wrapper
 
     return decorator
+
+
+def is_localhost_organizer(localhost_id: str, user=None) -> bool:
+    """
+    Check if user is a localhost organizer.
+    """
+    if not user:
+        user = frappe.session.user
+
+    # System Manager bypass
+    if "System Manager" in frappe.get_roles(user):
+        return True
+
+    profile = frappe.db.get_value(USER_PROFILE, {"user": user}, "name")
+    # Check if profile exists in localhost organizers
+    is_member = frappe.db.exists(
+        LOCALHOST_ORGANIZER,
+        {
+            "parent": localhost_id,
+            "parenttype": HACKATHON_LOCALHOST,
+            "profile": profile,
+            "parentfield": "organizers",
+        },
+    )
+
+    if not is_member:
+        frappe.throw(
+            "You are not a member of this Localhost. You are not authorized to view this page"
+        )
+        return False
+
+    return True
 
 
 def require_localhost_organizer(localhost_param="localhost_id"):
